+++ examples/anchor/drift/drift_interface/src/lib.rs
@@ -1,7 +1,9 @@
 solana_program::declare_id!("11111111111111111111111111111111");
 pub mod accounts;
 pub use accounts::*;
-pub mod instructions;
-pub use instructions::*;
 pub mod typedefs;
 pub use typedefs::*;
+pub mod instructions;
+pub use instructions::*;
+pub mod errors;
+pub use errors::*;