+++ examples/anchor/unstake-it/unstake_interface/src/lib.rs
@@ -1,7 +1,9 @@
 solana_program::declare_id!("unpXTU2Ndrc7WWNyEhQWe4udTzSibLPi25SXv2xbCHQ");
 pub mod accounts;
 pub use accounts::*;
-pub mod instructions;
-pub use instructions::*;
 pub mod typedefs;
 pub use typedefs::*;
+pub mod instructions;
+pub use instructions::*;
+pub mod errors;
+pub use errors::*;