+++ solores/src/idl_format/anchor/errors.rs
@@ -1,7 +1,8 @@
 use heck::ToPascalCase;
-use proc_macro2::TokenStream;
+use proc_macro2::{Span, TokenStream};
 use quote::{format_ident, quote, ToTokens};
 use serde::Deserialize;
+use syn::LitInt;
 
 #[derive(Deserialize)]
 pub struct ErrorEnumVariant {
@@ -14,10 +15,10 @@ impl ToTokens for ErrorEnumVariant {
     fn to_tokens(&self, tokens: &mut TokenStream) {
         let variant_ident = format_ident!("{}", self.name.to_pascal_case());
         let msg = &self.msg;
-        let code = self.code;
+        let code_literal = LitInt::new(&self.code.to_string(), Span::call_site());
         tokens.extend(quote! {
             #[error(#msg)]
-            #variant_ident = #code,
+            #variant_ident = #code_literal,
         })
     }
 }